PP-3676 Send email when creating mandate
- Added new email templates
- Hooked up to send emails for ON_DEMAND when mandate confirmed
- Split ON_DEMAND and ONE_OFF tests into separate files
