+++ src/main/java/uk/gov/pay/directdebit/notifications/model/EmailPayload.java
@@ -29,6 +29,8 @@ public class EmailPayload {
     }
 
     public enum EmailTemplate {
+        ON_DEMAND_MANDATE_CREATED,
+        ONE_OFF_MANDATE_CREATED,
         MANDATE_CANCELLED,
         MANDATE_FAILED,
         ONE_OFF_PAYMENT_CONFIRMED,
